third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/script-language-type-expected.txt

This is a testharness.js-based test.
FAIL A script with empty type should runassert_equals: expected true but got false
Harness: the test ran to completion.
